+++ src/nhs_reusable_code_library/standard_data_validations/nhsNumberValidation/sensitive_legally_restricted.py
@@ -1,0 +1,14 @@
+def sensitive_legally_restricted(decision_to_admit :str, nhs_number_status_indicator :str, witheld_identity_reason :str) -> bool:
+    """ This function checks if the NHS number is sensitive or legally restricted.
+    Args:
+    decision_to_admit (str): The decision to admit the patient.
+    nhs_number_status_indicator (str): The NHS number status indicator.
+    witheld_identity_reason (str): The reason for withholding the identity.
+    Returns:
+    bool: True if the NHS number is sensitive or legally restricted, False otherwise.
+    """
+    # here would go the code for determining sensitive/legally restricted, which is a common exclusion criteria in many CDS completeness rules. 
+    # It would take in the decision to admint, the nhs_number status indicator and the witheld identity reason, and output true if these indicate
+    # the data is sensitive / legally restricted.   
+
+    return True
